Vec to HashMap
