+++ src/envelopes.rs
@@ -67,15 +67,14 @@ impl Signal for SustainSynth {
     }
 
     fn signal(&mut self, graph: &Graph, sample_rate: Real) -> Real {
-        let amp = graph.output(self.wave) * self.calc_level();
+        let amp = graph.output(&self.wave) * self.calc_level();
         self.clock += 1. / sample_rate;
         self.level = self.calc_level();
         amp
     }
 }
 pub fn set_attack(graph: &Graph, n: Tag, a: Real) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -87,8 +86,7 @@ pub fn set_attack(graph: &Graph, n: Tag, a: Real) {
 }
 
 pub fn set_decay(graph: &Graph, n: Tag, d: Real) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -100,8 +98,7 @@ pub fn set_decay(graph: &Graph, n: Tag, d: Real) {
 }
 
 pub fn set_release(graph: &Graph, n: Tag, r: Real) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -113,8 +110,7 @@ pub fn set_release(graph: &Graph, n: Tag, r: Real) {
 }
 
 pub fn set_sustain_level(graph: &Graph, n: Tag, s: Real) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -126,8 +122,7 @@ pub fn set_sustain_level(graph: &Graph, n: Tag, s: Real) {
 }
 
 pub fn on(graph: &Graph, n: Tag) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -139,8 +134,7 @@ pub fn on(graph: &Graph, n: Tag) {
 }
 
 pub fn off(graph: &Graph, n: Tag) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
