+++ src/operators.rs
@@ -21,7 +21,7 @@ impl Signal for Product {
         self
     }
     fn signal(&mut self, graph: &Graph, _sample_rate: Real) -> Real {
-        self.waves.iter().fold(1.0, |acc, n| acc * graph.output(*n))
+        self.waves.iter().fold(1.0, |acc, n| acc * graph.output(n))
     }
 }
 
@@ -44,7 +44,7 @@ impl Signal for Mixer {
         self
     }
     fn signal(&mut self, graph: &Graph, _sample_rate: Real) -> Real {
-        self.waves.iter().fold(0.0, |acc, n| acc + graph.output(*n))
+        self.waves.iter().fold(0.0, |acc, n| acc + graph.output(n))
     }
 }
 
@@ -74,16 +74,15 @@ impl Signal for Lerp {
     }
 
     fn signal(&mut self, graph: &Graph, _sample_rate: Real) -> Real {
-        let alpha = In::val(graph, self.alpha);
-        alpha * In::val(graph, self.wave2) + (1.0 - alpha) * In::val(graph, self.wave1)
+        let alpha = In::val(graph, &self.alpha);
+        alpha * In::val(graph, &self.wave2) + (1.0 - alpha) * In::val(graph, &self.wave1)
     }
 }
 
 pub fn set_alpha(graph: &Graph, k: In, a: Real) {
     match k {
         In::Cv(n) => {
-            assert!(n < graph.nodes.len());
-            if let Some(v) = graph.nodes[n]
+            if let Some(v) = graph.nodes[&n]
                 .module
                 .lock()
                 .unwrap()
@@ -117,8 +116,8 @@ impl Lerp3 {
     }
 
     pub fn set_alphas(&mut self, graph: &Graph) {
-        let knob = In::val(graph, self.knob);
-        if In::val(graph, self.knob) <= 0.5 {
+        let knob = In::val(graph, &self.knob);
+        if In::val(graph, &self.knob) <= 0.5 {
             set_alpha(&graph, self.lerp1, 2.0 * knob);
             set_alpha(&graph, self.lerp2, 0.0);
         } else {
@@ -135,17 +134,16 @@ impl Signal for Lerp3 {
 
     fn signal(&mut self, graph: &Graph, _sample_rate: Real) -> Real {
         self.set_alphas(graph);
-        if In::val(graph, self.knob) <= 0.5 {
-            In::val(graph, self.lerp1)
+        if In::val(graph, &self.knob) <= 0.5 {
+            In::val(graph, &self.lerp1)
         } else {
-            In::val(graph, self.lerp2)
+            In::val(graph, &self.lerp2)
         }
     }
 }
 
 pub fn set_knob(graph: &Graph, n: Tag, k: Real) {
-    assert!(n < graph.nodes.len());
-    if let Some(v) = graph.nodes[n]
+    if let Some(v) = graph.nodes[&n]
         .module
         .lock()
         .unwrap()
@@ -185,10 +183,10 @@ impl Signal for Modulator {
     }
 
     fn signal(&mut self, graph: &Graph, _sample_rate: Real) -> Real {
-        let mod_hz = In::val(graph, self.mod_hz);
-        let mod_idx = In::val(graph, self.mod_idx);
-        let base_hz = In::val(graph, self.base_hz);
-        base_hz + mod_idx * mod_hz * In::val(graph, self.wave)
+        let mod_hz = In::val(graph, &self.mod_hz);
+        let mod_idx = In::val(graph, &self.mod_idx);
+        let base_hz = In::val(graph, &self.base_hz);
+        base_hz + mod_idx * mod_hz * In::val(graph, &self.wave)
     }
 }
 
@@ -220,8 +218,7 @@ impl IndexMut<&str> for Modulator {
 
 impl<'a> Set<'a> for Modulator {
     fn set(graph: &Graph, n: Tag, field: &str, value: Real) {
-        assert!(n < graph.nodes.len());
-        if let Some(v) = graph.nodes[n]
+        if let Some(v) = graph.nodes[&n]
             .module
             .lock()
             .unwrap()
